use version specific codeto load preloads uris
